zip-for-habrok/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cu128

accelerate==1.9.0
    # via
    #   -r requirements.in
    #   peft
    #   presidio-analyzer
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via fsspec
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via aiohttp
bitsandbytes==0.46.1
    # via -r requirements.in
blis==1.3.0
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.7.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via typer
cloudpathlib==0.21.1
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cryptography==44.0.3
    # via presidio-anonymizer
cymem==2.0.11
    # via
    #   preshed
    #   spacy
    #   thinc
datasets==4.0.0
    # via evaluate
dill==0.3.8
    # via
    #   datasets
    #   evaluate
    #   multiprocess
evaluate==0.4.5
    # via -r requirements.in
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
    #   tldextract
    #   torch
    #   transformers
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.3.0
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   torch
hf-xet==1.1.5
    # via huggingface-hub
huggingface-hub==0.33.4
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   evaluate
    #   peft
    #   presidio-analyzer
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   requests
    #   tldextract
    #   yarl
jinja2==3.1.6
    # via
    #   spacy
    #   torch
langcodes==3.5.0
    # via spacy
language-data==1.3.0
    # via langcodes
marisa-trie==1.2.1
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via
    #   datasets
    #   evaluate
murmurhash==1.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   accelerate
    #   bitsandbytes
    #   blis
    #   datasets
    #   evaluate
    #   pandas
    #   peft
    #   spacy
    #   thinc
    #   torchvision
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
packaging==25.0
    # via
    #   accelerate
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   peft
    #   spacy
    #   thinc
    #   transformers
    #   weasel
pandas==2.3.1
    # via
    #   datasets
    #   evaluate
peft==0.16.0
    # via -r requirements.in
phonenumbers==9.0.10
    # via presidio-analyzer
pillow==11.3.0
    # via torchvision
preshed==3.0.10
    # via
    #   spacy
    #   thinc
presidio-analyzer[transformers]==2.2.359
    # via -r requirements.in
presidio-anonymizer==2.2.359
    # via -r requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via
    #   accelerate
    #   peft
pyarrow==21.0.0
    # via datasets
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   peft
    #   presidio-analyzer
    #   transformers
regex==2024.11.6
    # via
    #   presidio-analyzer
    #   transformers
requests==2.32.4
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   requests-file
    #   spacy
    #   tldextract
    #   transformers
    #   weasel
requests-file==2.1.0
    # via tldextract
rich==14.0.0
    # via typer
safetensors==0.5.3
    # via
    #   accelerate
    #   peft
    #   transformers
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smart-open==7.3.0.post1
    # via weasel
spacy==3.8.7
    # via
    #   presidio-analyzer
    #   spacy-huggingface-pipelines
spacy-huggingface-pipelines==0.0.4
    # via presidio-analyzer
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.1
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
sympy==1.14.0
    # via torch
thinc==8.3.6
    # via spacy
tldextract==5.3.0
    # via presidio-analyzer
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   accelerate
    #   bitsandbytes
    #   peft
    #   spacy-huggingface-pipelines
    #   torchvision
torchvision==0.22.1+cu128
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   datasets
    #   evaluate
    #   huggingface-hub
    #   peft
    #   spacy
    #   transformers
transformers==4.53.3
    # via
    #   -r requirements.in
    #   peft
    #   presidio-analyzer
    #   spacy-huggingface-pipelines
triton==3.3.1
    # via torch
typer==0.16.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
wrapt==1.17.2
    # via smart-open
xxhash==3.5.0
    # via
    #   datasets
    #   evaluate
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
